Fix to #13642 - Comparison of nullable result of functions
Problem was that we were not applying clr null semantics when comparing result of functions, i.e. if the function results in null and we compared it to another null value, we would treat them as non-equal. According to clr null semantics they should be treated as the same.

Fix is to expand IsNullExpression builder to handle functions with nullable arguments.

f(a, b) == null <=> a == null || b == null, unless function is COALESCE (in that case both arguments have to be null), so we can just test arguments not the entire function expression.
Also added small optimization in PredicateReductionExpressionOptimizer that removes redundant IS NULL terms.